backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
pillow
numpy
google-generativeai
python-dotenv
icrawler
pandas
xlrd
